+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.8.1
+* feat: add support for downloading models from modelscope
 * fix: fix list index out of range error caused by calling LayoutElements.from_list() with empty list
 
 ## 0.8.0